+++ CHANGELOG.md
@@ -50,7 +50,7 @@
 * `SBBAccordion`: Added field `titleMaxLines` (with default value `null`, meaning titles are now multiline by default)
 * `SBBAccordion`: Added constructor `single` for simpler usage when only one item is needed
 * `SBBListHeader`: Added field `maxLines` (with default value `null`, meaning list headers are now multiline by default)
-* `SBBTabBar`: Added field `onTap`, which allows for reacting to subsequent taps on the selected tab item.
+* `SBBTabBar`: Added field `onTap`, which allows for reacting to taps on tab items.
 
 ### Changed